refactor(modal): extract owner read helper boundary
